Track IDs of created objects across frontend update runs.
Until the sync handler updates and writes a new view file based on
Concertim data, we need to be able to keep track of those Concertim
objects that we've created - otherwise we'll immediately attempt to
create them again on the subsequent run. Unless we save the IDs of
objects we create in Concertim into the view, we'll keep trying (and
failing) to create them again until the sync handler runs.

The current logic is that the sync handler will blow away any changes we
make to this file; but that's fine because the new data from the sync
handler will contain the IDs we need anyway.

The one remaining concern here is that `sync` and `fe_updates` runs happen
simultaneously - which is fine unless the `fe_updates` finishes later,
in which case the new Concertim data in sync's view will be overwritten
by the incomplete data from `fe_updates`. Two possible solutions come to
mind:

1. Update the logic of `merge_views` so that `sync`'s view always
   overwrites a view from `fe_updates`, regardless of timestamp;
2. Prevent `sync` from writing a view while `fe_updates` is running.

Or, alternatively, ensure that `sync` and `fe_updates` can't run in
parallel by always running the latter immediately after the former in
the same process.