Adding 12:00:00 default time to each record_usage Killbill call for daily usage updates
Killbill record_usage API call assumes time to be 00:00:00 by default, which causes the usage to not show up in invoices, if the usage date and subscription start date is the same. record_usage call should be ideally used once per day. Adding a default time of 12:00:00 fixes this issue.